pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "beautifulsoup4>=4.13.5",
    "discord-py>=2.5.2",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.1",
    "google-genai>=1.27.0",
    "gunicorn>=23.0.0",
    "openai>=1.97.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "qrcode>=8.2",
    "requests>=2.32.4",
    "sift-stack-py>=0.7.0",
    "yt-dlp>=2025.9.5",
    "sqlalchemy>=2.0.43",
    "pynacl>=1.6.0",
]
